tratativa de erros

+++ api/src/agent/service/agent.service.ts
@@ -1,5 +1,5 @@
 import { AgentRepository } from "../repository/agent.repository";
-import { Injectable } from "@nestjs/common";
+import { BadRequestException, ConflictException, Injectable } from "@nestjs/common";
 import { AgentRequest } from "../integration/request/agent.request";
 import { AgentMapper } from "../mapper/agent.mapper";
 import { AgentResponse } from "../integration/response/agent.response";
@@ -8,13 +8,17 @@ import { ImageRequest } from "src/common/integration/request/image.request";
 import { Builder } from "builder-pattern";
 import { ImageSizeEnum } from "src/common/enum/image-size.enum";
 import { Response } from "express";
+import { PropertyResponse } from "src/property/property/integration/response/property.response";
+import { PropertyRepository } from "src/property/property/repository/property.repository";
+import { PropertyMapper } from "src/property/property/mapper/property.mapper";
 
 @Injectable()
 export class AgentService {
   constructor(
     private repository: AgentRepository,
-    private imageService: ImageService
-  ) {}
+    private imageService: ImageService,
+    private propertyRepository: PropertyRepository
+  ) { }
 
   public insert(request: AgentRequest): Promise<number> {
     return this.repository.insert(AgentMapper.requestToEntity(request));
@@ -46,8 +50,28 @@ export class AgentService {
     return this.repository.update(code, AgentMapper.requestToEntity(request));
   }
 
-  public delete(code: number): Promise<number> {
-    return this.repository.delete(code);
+  public getProperties(code: number): Promise<PropertyResponse[]> {
+    return this.propertyRepository.getAll({ paginacao: { pagina: 1, porPagina: 1000 }, agenciador: code })
+      .then(result => PropertyMapper.entityListToResponse(result));
+  }
+
+  public async delete(code: number): Promise<number> {
+    const properties = await this.getProperties(code);
+    console.log({ properties })
+
+    if ((properties && properties.length)) {
+      const message = 'Não foi possivel deletar o registro pois o mesmo contem outros registros vinculados.';
+      throw new BadRequestException({
+        message,
+        properties,
+      }, message);
+    } else {
+      try {
+        return this.repository.delete(code);
+      } catch (error) {
+        throw new ConflictException(error, 'Erro ao executar comando no banco de dados')
+      }
+    }
   }
 
   public buildAgentImage(file: Express.Multer.File): ImageRequest {
